- Deffered Shading 관련된 로직 구현 완료 - 수정 필요(Deffered 수행하지 않는 부분과 결합 등)

+++ DirectX11/IndexBuffer.h
@@ -23,7 +23,7 @@ protected:
 template<typename T>
 inline void IndexBuffer<T>::InitializeForGPU(const UINT& ArrayCountIn, T* CPUDataPtr)
 {
-	IndicesCount = static_cast<UINT>(ArrayCountIn);
+	IndicesCount = ArrayCountIn;
 
 	D3D11_BUFFER_DESC BufferDesc;
 	AutoZeroMemory(BufferDesc);
